Refactor 1 css stylesheet into 1 stylesheet/component
