+++ client/src/components/carousel.jsx
@@ -3,7 +3,7 @@ import CarouselEntry from './CarouselEntry.jsx';
 import ForwardButton from './ForwardButton.jsx';
 import BackButton from './BackButton.jsx';
 import FavoritesModal from './FavoritesModal.jsx';
-import styles from './../styles/style.css';
+import styles from './../styles/carousel.css';
 
 const Carousel = ({ currentListings, scrollCarousel, lastThreeListings, firstThreeListings, modalIsShown, showModal, closeModal }) => {
   const modal = modalIsShown ? <FavoritesModal closeModal={closeModal}/> : null;
