+++ client/src/styles/app.css
@@ -1,0 +1,4 @@
+.serviceContainer {
+  display: flex;
+  justify-content: center;
+}
