+++ client/src/styles/carousel-entry.css
@@ -1,0 +1,68 @@
+.photoContainer {
+  position: relative;
+  margin: 0 8px 0 8px;
+  width: 330px;
+}
+
+.photo {
+  width: 330px;
+  height: 230px;
+  border-radius: 3px;
+  /* cursor: pointer; */
+}
+
+.heartButton {
+  width: 25px;
+  height: 25px;
+  position: absolute;
+  top: 10px;
+  left: 295px;
+  background-color: transparent;
+  border: 0;
+  padding: 0;
+  cursor: pointer;
+}
+
+.heart {
+  width: 25px;
+  height: 25px;
+}
+
+.descriptionAndBeds {
+  margin-top: 10px;
+  margin-bottom: 5px;
+  font-weight: bold;
+  font-size: 12px;
+  color: #0b6054;
+}
+
+.header {
+  margin-top: 5px;
+  margin-bottom: 5px;
+  font-weight: bold;
+  color: #424447;
+}
+
+.price {
+  margin-top: 5px;
+  margin-bottom: 5px;
+  font-weight: lighter;
+  font-size: 14px;
+  color: #5d5d60;
+}
+
+.starsAndReviews {
+  margin-top: 5px;
+  margin-bottom: 5px;
+  font-size: 12px;
+  font-weight: bold;
+  color: #4b4b4f;
+}
+
+.stars {
+  display: inline;
+}
+
+.reviews {
+  margin: 0 3px;
+}
