--- client/src/styles/style.css
@@ -1,164 +1,0 @@
-.serviceContainer {
-  display: flex;
-  justify-content: center;
-}
-
-.carousel {
-  display: flex;
-  justify-content: center;
-}
-
-.listingsHeader {
-  color: #424447;
-  margin: 19px 0 19px 32px;
-}
-
-.photoContainer {
-  position: relative;
-  margin: 0 8px 0 8px;
-  width: 330px;
-}
-
-.photo {
-  width: 330px;
-  height: 230px;
-  border-radius: 3px;
-  /* cursor: pointer; */
-}
-
-.heartButton {
-  width: 25px;
-  height: 25px;
-  position: absolute;
-  top: 10px;
-  left: 295px;
-  background-color: transparent;
-  border: 0;
-  padding: 0;
-  cursor: pointer;
-}
-
-.heart {
-  width: 25px;
-  height: 25px;
-}
-
-.descriptionAndBeds {
-  margin-top: 10px;
-  margin-bottom: 5px;
-  font-weight: bold;
-  font-size: 12px;
-  color: #0b6054;
-}
-
-.header {
-  margin-top: 5px;
-  margin-bottom: 5px;
-  font-weight: bold;
-  color: #424447;
-}
-
-.price {
-  margin-top: 5px;
-  margin-bottom: 5px;
-  font-weight: lighter;
-  font-size: 14px;
-  color: #5d5d60;
-}
-
-.starsAndReviews {
-  margin-top: 5px;
-  margin-bottom: 5px;
-  font-size: 12px;
-  font-weight: bold;
-  color: #4b4b4f;
-}
-
-.stars {
-  display: inline;
-}
-
-.reviews {
-  margin: 0 3px;
-}
-
-.buttonContainer {
-  display: flex;
-  height: 230px;
-  align-items: center;
-  width: 25px;
-}
-
-.scrollButton {
-  border: 0px;
-  padding: 0px;
-  height: 25px;
-  cursor: pointer;
-}
-
-.img {
-  width: 25px;
-  height: 25px;
-}
-
-.modalBackground {
-  position: fixed;
-  z-index: 1;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  overflow: auto;
-  background-color: rgba(0,0,0,0.7);
-}
-
-.modalContent {
-  background-color: #fefefe;
-  margin: 65px auto;
-  padding: 32px;
-  width: 504px;
-  height: 609px;
-}
-
-.closeButton {
-  background-color: transparent;
-  border: none;
-  width: 16px;
-  height: 16px;
-  padding: 0;
-  cursor: pointer;
-  border-radius: 50%;
-}
-
-.closeImg {
-  width: 16px;
-  height: 16px;
-}
-
-.modalHeader {
-  color: #424447;
-}
-
-.facebookButton {
-  width: 100%;
-  height: 49px;
-  border: 2px solid #3B5998;
-  border-radius: 4px;
-  padding: 12px 24px;
-  background-color: #3B5998;
-  color: white;
-  margin-bottom: 8px;
-}
-
-.googleButton {
-  width: 100%;
-  height: 49px;
-  border: 2px solid #545a63;
-  border-radius: 4px;
-  padding: 12px 24px;
-}
-
-.facebookImg {
-  width: 25px;
-  height: 25px;
-}
